% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_changer_data.R
\docType{data}
\name{build_changer_data}
\alias{build_changer_data}
\title{Builder for changer analysis, comparing two years of prescribing and payments}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
build_changer_data
}
\description{
Builder for changer analysis, comparing two years of prescribing and payments
}
\keyword{datasets}

