% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_drug_class.R
\docType{data}
\name{build_drug_class}
\alias{build_drug_class}
\title{Builder Class for Drug Class Study Groups}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
build_drug_class
}
\description{
Builder Class for Drug Class Study Groups
}
\keyword{datasets}

