% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_figures.R
\docType{data}
\name{build_figures}
\alias{build_figures}
\title{Builder class for figures}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
build_figures
}
\description{
Builder class for figures
}
\keyword{datasets}

