% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\docType{data}
\name{build_model}
\alias{build_model}
\title{Analysis class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
build_model
}
\description{
Analysis class
}
\keyword{datasets}

