% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_openpay_year.R
\docType{data}
\name{build_openpay_year}
\alias{build_openpay_year}
\title{Builder Class for Open Payments data by year}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
build_openpay_year
}
\description{
Builder Class for Open Payments data by year
}
\keyword{datasets}

