% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prescriber_year.R
\docType{data}
\name{build_prescriber_year}
\alias{build_prescriber_year}
\title{Builder Class for Part D medicare prescribing data by year}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
build_prescriber_year
}
\description{
Builder Class for Part D medicare prescribing data by year
}
\keyword{datasets}

