% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_study_group.R
\docType{data}
\name{build_study_group}
\alias{build_study_group}
\title{Builder Class for Drug Class Study Groups}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
build_study_group
}
\description{
Builder Class for Drug Class Study Groups
}
\keyword{datasets}

