% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_pop_library.R
\docType{data}
\name{study_pop_library}
\alias{study_pop_library}
\title{Library Class to hold exlusion criteria, study population numbers, and other descriptive data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
study_pop_library
}
\description{
Library Class to hold exlusion criteria, study population numbers, and other descriptive data
}
\keyword{datasets}

